Switch to using longs, not ints
git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222701 13f79535-47bb-0310-9956-ffa450edef68

+++ security/src/java/org/apache/fulcrum/security/entity/SecurityEntity.java
@@ -87,12 +87,12 @@ public interface SecurityEntity
      *
      * @return The Id of the SecurityEntity.
      */
-    int getId();
+    long getId();
 
     /**
      * Sets the Id of the SecurityEntity.
      *
      * @param id The new Id of the SecurityEntity
      */
-    void setId(int id);
+    void setId(long id);
 }
